REF: CategoricalIndex indexing tests (#31559)
